refact(publish_filters) - Renamed the services
