+++ server/apps/publish/publish_filters/filter_condition_tests.py
@@ -9,7 +9,7 @@
 # at https://www.sourcefabric.org/superdesk/license
 
 from superdesk.tests import TestCase
-from apps.publish.blocking_filters.filter_condition import FilterConditionService
+from apps.publish.publish_filters.filter_condition import FilterConditionService
 from eve.utils import ParsedRequest
 import json
 import superdesk
@@ -47,7 +47,7 @@ class FilterConditionTests(TestCase):
         f._translate_to_mongo_query(doc)
         with self.app.app_context():
             docs = superdesk.get_resource_service('archive').\
-                get_from_mongo(req=self.req, lookup=doc['mongo_translation'])
+                get_from_mongo(req=self.req, lookup=doc['mongo_query'])
             self.assertEquals(1, docs.count())
             self.assertEquals('1', docs[0]['_id'])
 
@@ -57,7 +57,7 @@ class FilterConditionTests(TestCase):
         f._translate_to_mongo_query(doc)
         with self.app.app_context():
             docs = superdesk.get_resource_service('archive').\
-                get_from_mongo(req=self.req, lookup=doc['mongo_translation'])
+                get_from_mongo(req=self.req, lookup=doc['mongo_query'])
             doc_ids = [doc['_id'] for doc in docs]
             self.assertEquals(2, docs.count())
             self.assertTrue('1' in doc_ids)
@@ -69,7 +69,7 @@ class FilterConditionTests(TestCase):
         f._translate_to_mongo_query(doc)
         with self.app.app_context():
             docs = superdesk.get_resource_service('archive').\
-                get_from_mongo(req=self.req, lookup=doc['mongo_translation'])
+                get_from_mongo(req=self.req, lookup=doc['mongo_query'])
             self.assertEquals(1, docs.count())
             self.assertEquals('1', docs[0]['_id'])
 
@@ -79,7 +79,7 @@ class FilterConditionTests(TestCase):
         f._translate_to_mongo_query(doc)
         with self.app.app_context():
             docs = superdesk.get_resource_service('archive').\
-                get_from_mongo(req=self.req, lookup=doc['mongo_translation'])
+                get_from_mongo(req=self.req, lookup=doc['mongo_query'])
             self.assertEquals(1, docs.count())
             self.assertEquals('2', docs[0]['_id'])
 
@@ -89,7 +89,7 @@ class FilterConditionTests(TestCase):
         f._translate_to_mongo_query(doc)
         with self.app.app_context():
             docs = superdesk.get_resource_service('archive').\
-                get_from_mongo(req=self.req, lookup=doc['mongo_translation'])
+                get_from_mongo(req=self.req, lookup=doc['mongo_query'])
             self.assertEquals(5, docs.count())
             doc_ids = [doc['_id'] for doc in docs]
             self.assertTrue('2' not in doc_ids)
@@ -100,7 +100,7 @@ class FilterConditionTests(TestCase):
         f._translate_to_mongo_query(doc)
         with self.app.app_context():
             docs = superdesk.get_resource_service('archive').\
-                get_from_mongo(req=self.req, lookup=doc['mongo_translation'])
+                get_from_mongo(req=self.req, lookup=doc['mongo_query'])
             self.assertEquals(2, docs.count())
             self.assertEquals('3', docs[0]['_id'])
             self.assertEquals('4', docs[1]['_id'])
@@ -111,7 +111,7 @@ class FilterConditionTests(TestCase):
         f._translate_to_mongo_query(doc)
         with self.app.app_context():
             docs = superdesk.get_resource_service('archive').\
-                get_from_mongo(req=self.req, lookup=doc['mongo_translation'])
+                get_from_mongo(req=self.req, lookup=doc['mongo_query'])
             self.assertEquals(3, docs.count())
             doc_ids = [doc['_id'] for doc in docs]
             self.assertTrue('1', doc_ids)
@@ -122,7 +122,7 @@ class FilterConditionTests(TestCase):
         doc = {'field': 'urgency', 'operator': 'in', 'value': '3,4'}
         f._translate_to_elastic_query(doc)
         with self.app.app_context():
-            self._setup_elastic_args(doc['elastic_translation'])
+            self._setup_elastic_args(doc['elastic_query'])
             docs = superdesk.get_resource_service('archive').get(req=self.req, lookup=None)
             doc_ids = [doc['_id'] for doc in docs]
             self.assertEquals(2, docs.count())
@@ -134,7 +134,7 @@ class FilterConditionTests(TestCase):
         doc = {'field': 'urgency', 'operator': 'nin', 'value': '3,4'}
         f._translate_to_elastic_query(doc)
         with self.app.app_context():
-            self._setup_elastic_args(doc['elastic_translation'], 'not')
+            self._setup_elastic_args(doc['elastic_query'], 'not')
             docs = superdesk.get_resource_service('archive').get(req=self.req, lookup=None)
             self.assertEquals(4, docs.count())
             doc_ids = [doc['_id'] for doc in docs]
@@ -146,7 +146,7 @@ class FilterConditionTests(TestCase):
         doc = {'field': 'headline', 'operator': 'like', 'value': 'Tor'}
         f._translate_to_elastic_query(doc)
         with self.app.app_context():
-            self._setup_elastic_args(doc['elastic_translation'], 'keyword')
+            self._setup_elastic_args(doc['elastic_query'], 'keyword')
             docs = superdesk.get_resource_service('archive').get(req=self.req, lookup=None)
             self.assertEquals(2, docs.count())
             doc_ids = [doc['_id'] for doc in docs]
@@ -158,7 +158,7 @@ class FilterConditionTests(TestCase):
         doc = {'field': 'headline', 'operator': 'notlike', 'value': 'que'}
         f._translate_to_elastic_query(doc)
         with self.app.app_context():
-            self._setup_elastic_args(doc['elastic_translation'], 'not')
+            self._setup_elastic_args(doc['elastic_query'], 'not')
             docs = superdesk.get_resource_service('archive').get(req=self.req, lookup=None)
             self.assertEquals(5, docs.count())
             doc_ids = [doc['_id'] for doc in docs]
@@ -169,7 +169,7 @@ class FilterConditionTests(TestCase):
         doc = {'field': 'headline', 'operator': 'startswith', 'value': 'Sto'}
         f._translate_to_elastic_query(doc)
         with self.app.app_context():
-            self._setup_elastic_args(doc['elastic_translation'], 'keyword')
+            self._setup_elastic_args(doc['elastic_query'], 'keyword')
             docs = superdesk.get_resource_service('archive').get(req=self.req, lookup=None)
             self.assertEquals(1, docs.count())
             self.assertEquals('1', docs[0]['_id'])
@@ -179,7 +179,7 @@ class FilterConditionTests(TestCase):
         doc = {'field': 'headline', 'operator': 'endswith', 'value': 'Que'}
         f._translate_to_elastic_query(doc)
         with self.app.app_context():
-            self._setup_elastic_args(doc['elastic_translation'], 'keyword')
+            self._setup_elastic_args(doc['elastic_query'], 'keyword')
             docs = superdesk.get_resource_service('archive').get(req=self.req, lookup=None)
             self.assertEquals(1, docs.count())
             self.assertEquals('2', docs[0]['_id'])
