+++ server/settings.py
@@ -172,7 +172,7 @@ INSTALLED_APPS = [
     'apps.rules',
     'apps.highlights',
     'apps.publish',
-    'apps.publish.blocking_filters',
+    'apps.publish.publish_filters',
     'apps.macros',
     'apps.dictionaries',
     'apps.duplication',
